Check admin classes not in admin.py
This is because registering an admin class implies a call to
models.get_apps() which attempts to import *every* models.py in
the project.

Whilst your project should probably not have significant
inter-app dependencies, importing every possible models.py does
not help the situation and can cause ImportError when models are
loaded in different scenarios.

Signed-off-by: Chris Lamb <lamby@debian.org>