Apply patch from teh:
updating pylint to the squeeze version breaks "django-lint -e" for me.
There are two issues:

1) The deprecated option disable-msg-cat was removed
2) You can only disable categories that have been registered before

The attached patch fixes that for me, but not in a very elegant way. I
think the real fixes need to go to pylint, but I don't want to spend
time on this right now.


Signed-off-by: Chris Lamb <lamby@debian.org>